+++ crypto/src/commonMain/kotlin/dev/inmo/micro_utils/crypto/Base64.kt
@@ -1,4 +1,4 @@
-package dev.inmo.micro_utils.common
+package dev.inmo.micro_utils.crypto
 
 import kotlin.experimental.and
 
@@ -12,20 +12,13 @@ private val BASE64_INVERSE_ALPHABET = IntArray(256) {
 internal fun Int.toBase64(): Char = BASE64_ALPHABET[this]
 internal fun Byte.fromBase64(): Byte = BASE64_INVERSE_ALPHABET[toInt() and 0xff].toByte() and BASE64_MASK
 
-const val replacedInCrypto = "Replaced in crypto subproject"
-
-@Deprecated(replacedInCrypto, ReplaceWith("EncodedBase64String", "dev.inmo.micro_utils.crypto"))
 typealias EncodedBase64String = String
-@Deprecated(replacedInCrypto, ReplaceWith("EncodedByteArray", "dev.inmo.micro_utils.crypto"))
 typealias EncodedByteArray = ByteArray
 
-@Deprecated(replacedInCrypto, ReplaceWith("encodeBase64String()", "dev.inmo.micro_utils.crypto"))
-fun String.encodeBase64String(): EncodedBase64String = encodeToByteArray().encodeBase64String()
-@Deprecated(replacedInCrypto, ReplaceWith("encodeBase64()", "dev.inmo.micro_utils.crypto"))
-fun String.encodeBase64(): EncodedByteArray = encodeToByteArray().encodeBase64()
+fun SourceString.encodeBase64String(): EncodedBase64String = encodeToByteArray().encodeBase64String()
+fun SourceString.encodeBase64(): EncodedByteArray = encodeToByteArray().encodeBase64()
 
-@Deprecated(replacedInCrypto, ReplaceWith("encodeBase64String()", "dev.inmo.micro_utils.crypto"))
-fun ByteArray.encodeBase64String(): EncodedBase64String = buildString {
+fun SourceBytes.encodeBase64String(): EncodedBase64String = buildString {
     var i = 0
     while (this@encodeBase64String.size > i) {
         val read = kotlin.math.min(3, this@encodeBase64String.size - i)
@@ -52,16 +45,12 @@ fun ByteArray.encodeBase64String(): EncodedBase64String = buildString {
         i += read
     }
 }
-@Deprecated(replacedInCrypto, ReplaceWith("encodeBase64()", "dev.inmo.micro_utils.crypto"))
-fun ByteArray.encodeBase64(): EncodedByteArray = encodeBase64String().encodeToByteArray()
+fun SourceBytes.encodeBase64(): EncodedByteArray = encodeBase64String().encodeToByteArray()
 
-@Deprecated(replacedInCrypto, ReplaceWith("decodeBase64()", "dev.inmo.micro_utils.crypto"))
-fun EncodedBase64String.decodeBase64() = dropLastWhile { it == BASE64_PAD }.encodeToByteArray().decodeBase64()
-@Deprecated(replacedInCrypto, ReplaceWith("decodeBase64String()", "dev.inmo.micro_utils.crypto"))
-fun EncodedBase64String.decodeBase64String() = decodeBase64().decodeToString()
+fun EncodedBase64String.decodeBase64(): SourceBytes = dropLastWhile { it == BASE64_PAD }.encodeToByteArray().decodeBase64()
+fun EncodedBase64String.decodeBase64String(): SourceString = decodeBase64().decodeToString()
 
-@Deprecated(replacedInCrypto, ReplaceWith("decodeBase64()", "dev.inmo.micro_utils.crypto"))
-fun EncodedByteArray.decodeBase64(): ByteArray {
+fun EncodedByteArray.decodeBase64(): SourceBytes {
     val result = mutableListOf<Byte>()
     val data = ByteArray(4)
     (0 until size step 4).forEach { i ->
@@ -85,5 +74,4 @@ fun EncodedByteArray.decodeBase64(): ByteArray {
 
     return result.toByteArray()
 }
-@Deprecated(replacedInCrypto, ReplaceWith("decodeBase64String()", "dev.inmo.micro_utils.crypto"))
-fun EncodedByteArray.decodeBase64String() = decodeBase64().decodeToString()
+fun EncodedByteArray.decodeBase64String(): SourceString = decodeBase64().decodeToString()