+++ crypto/src/commonTest/kotlin/dev/inmo/micro_utils/crypto/Base64Text.kt
@@ -1,6 +1,5 @@
-package dev.inmo.micro_utils.common
+package dev.inmo.micro_utils.crypto
 
-import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals